third_party/blink/web_tests/external/wpt/editing/run/fontname_1001-2000-expected.txt

This is a testharness.js-based test.
Found 64 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "[foo<pre>bar</pre>baz]" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "[foo<samp>bar</samp>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-family:sans-serif\\">foo<samp><span style=\\"font-family:sans-serif\\">bar</span></samp>baz</span>" but got "<span style=\\"font-family:sans-serif\\">foo<samp>bar</samp>baz</span>"
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "[foo<samp>bar</samp>baz]" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "[foo<samp>bar</samp>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "[foo<samp>bar</samp>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<font face=\\"sans-serif\\">foo<samp><font face=\\"sans-serif\\">bar</font></samp>baz</font>" but got "<font face=\\"sans-serif\\">foo<samp>bar</samp>baz</font>"
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "[foo<samp>bar</samp>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "[foo<tt>bar</tt>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<span style=\\"font-family:sans-serif\\">foo<tt><span style=\\"font-family:sans-serif\\">bar</span></tt>baz</span>" but got "<span style=\\"font-family:sans-serif\\">foo<tt>bar</tt>baz</span>"
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "[foo<tt>bar</tt>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "[foo<tt>bar</tt>baz]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<font face=\\"sans-serif\\">foo<tt><font face=\\"sans-serif\\">bar</font></tt>baz</font>" but got "<font face=\\"sans-serif\\">foo<tt>bar</tt>baz</font>"
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "[foo<tt>bar</tt>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "[foo<code>ba]r</code>baz" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "[foo<code>ba]r</code>baz" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "[foo<kbd>ba]r</kbd>baz" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "[foo<kbd>ba]r</kbd>baz" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "[foo<listing>ba]r</listing>baz" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "[foo<listing>ba]r</listing>baz" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "[foo<pre>ba]r</pre>baz" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "[foo<pre>ba]r</pre>baz" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "[foo<samp>ba]r</samp>baz" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "[foo<samp>ba]r</samp>baz" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "[foo<tt>ba]r</tt>baz" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "[foo<tt>ba]r</tt>baz" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "foo<code>b[ar</code>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "foo<code>b[ar</code>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "foo<kbd>b[ar</kbd>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "foo<kbd>b[ar</kbd>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "foo<listing>b[ar</listing>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "foo<listing>b[ar</listing>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "foo<pre>b[ar</pre>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "foo<pre>b[ar</pre>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "foo<samp>b[ar</samp>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "foo<samp>b[ar</samp>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "foo<tt>b[ar</tt>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "foo<tt>b[ar</tt>baz]" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["fontname","sans-serif"]] "foo<span style=\\"font-family: sans-serif\\">b[a]r</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<span style=\\"font-family:sans-serif\\">bar</span>baz" but got "foo<span style=\\"font-family:sans-serif\\">b</span><font face=\\"sans-serif\\">a</font><span style=\\"font-family:sans-serif\\">r</span>baz"
[FAIL] [["fontname","sans-serif"]] "foo<tt contenteditable=false>ba[r</tt>b]az" queryCommandValue("fontname") before
  assert_equals: Wrong result returned expected "serif" but got "monospace"
[FAIL] [["fontname","sans-serif"]] "foo<tt contenteditable=false>ba[r</tt>b]az" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "serif" but got "monospace"
[FAIL] [["fontname","sans-serif"]] "fo[o<tt contenteditable=false>b]ar</tt>baz": execCommand("fontname", false, "sans-serif") return value
  assert_equals: expected false but got true
[FAIL] [["fontname","sans-serif"]] "fo[o<tt contenteditable=false>b]ar</tt>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<tt contenteditable=\\"false\\">bar</tt>baz" but got "fo<font face=\\"sans-serif\\">o</font><tt contenteditable=\\"false\\">bar</tt>baz"
[FAIL] [["fontname","sans-serif"]] "fo[o<tt contenteditable=false>b]ar</tt>baz" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "serif" but got "sans-serif"
[FAIL] [["fontname","sans-serif"]] "foo<tt>{}<br></tt>bar" queryCommandValue("fontname") before
  assert_equals: Wrong result returned expected "monospace" but got "serif"
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "foo<tt>{<br></tt>}bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<tt><span style=\\"font-family:sans-serif\\"><br></span></tt>bar" but got "foo<tt style=\\"font-family:sans-serif\\"><br></tt>bar"
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "foo<tt>{<br></tt>}bar" queryCommandValue("fontname") before
  assert_equals: Wrong result returned expected "monospace" but got "serif"
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "foo<tt>{<br></tt>}bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "foo<tt>{<br></tt>}bar" queryCommandValue("fontname") before
  assert_equals: Wrong result returned expected "monospace" but got "serif"
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "foo<tt>{<br></tt>}bar" queryCommandValue("fontname") after
  assert_equals: Wrong result returned expected "sans-serif" but got "serif"
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "foo<tt>{<br></tt>b]ar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<tt><span style=\\"font-family:sans-serif\\"><br></span></tt><span style=\\"font-family:sans-serif\\">b</span>ar" but got "foo<span style=\\"font-family:sans-serif\\"><tt><br></tt>b</span>ar"
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "foo<tt>{<br></tt>b]ar" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "foo<tt>{<br></tt>b]ar" queryCommandValue("fontname") before
  assert_equals: Wrong result returned expected "monospace" but got "serif"
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "foo<tt>{<br></tt>b]ar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<tt><font face=\\"sans-serif\\"><br></font></tt><font face=\\"sans-serif\\">b</font>ar" but got "foo<font face=\\"sans-serif\\"><tt><br></tt>b</font>ar"
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "foo<tt>{<br></tt>b]ar" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "foo<tt>{<br></tt>b]ar" queryCommandValue("fontname") before
  assert_equals: Wrong result returned expected "monospace" but got "serif"
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "fo[o<span style=font-family:monospace>b]ar</span>baz" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "fo[o<span style=font-family:monospace>b]ar</span>baz" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "foo<span style=font-family:monospace>ba[r</span>b]az" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "foo<span style=font-family:monospace>ba[r</span>b]az" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "fo[o<span style=font-family:monospace>bar</span>b]az" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "fo[o<span style=font-family:monospace>bar</span>b]az" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "fo[o<code>b]ar</code>" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "fo[o<code>b]ar</code>" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "fo[o<kbd>b]ar</kbd>" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "fo[o<kbd>b]ar</kbd>" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["fontname","sans-serif"]] "fo[o<listing>b]ar</listing>" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["fontname","sans-serif"]] "fo[o<listing>b]ar</listing>" queryCommandIndeterm("fontname") before
  assert_equals: Wrong result returned expected true but got false
Harness: the test ran to completion.
